Add files via upload
This is the source code for a pricing page for a music streaming app called GrooveVerse. In this, You an see three plans containing features and the price per month. The unique part of it is that it contains a drop down menu containing some country names and when selected, display the price in that selected country's currency.